requirements.txt
python-dotenv
rich
cryptography
flask
flask-sqlalchemy
flask-bcrypt
flask-wtf
flask-login
flask-session
flask-debugtoolbar
requests